move bid to action file
